Fix webhook cache to not allow multiple entries per key
When many client requests with identical client ids arrive, it could happen
that multiple requests are processed in parallel. Nevertheless, those
must not lead to multiple cache entries per key, as this would crash
the lookup function with the below error message. Changing the ets storage
table from a `bag` to a `set` is apparently already sufficient to achieve
this.

Error message before this change:

13:02:18.953 [error] CRASH REPORT Process <0.2338.0> with 0 neighbours crashed with reason: no case clause matching [{{<<"https://auth-server:443/api/broker/auth/register">>,auth_on_register,[{addr,<<"172.21.0.7">>},{mountpoint,<<"password">>},{client_id,<<"client11">>},{username,<<"admin">>},{password,<<"admin">>},{clean_session,true}]},{<<"password">>,<<"client11">>},1652447234,[{mountpoint,<<>>},{client_id,<<"client11">>}]},{{<<"https://auth-server:443/api/broker/auth/register">>,auth_on_register,[{addr,<<"172.21.0.7">>},{mountpoint,<<"password">>},{client_id,<<"client11">>},{username,<<"admin">>},{password,...},...]},...}] in vmq_webhooks_cache:lookup_/3 line 88